fix:		fix SLF4J logging dependencies 	affected files:
	app/build.gradle
	* add SLF4J v 1.7.25 dependencies

refactor:  	cnage types of StockDao/IStockDap, add tests, add permissions

	affected files:

	app/src/test/java/com/example/devbox/stockhawkrewrite/UtilTest.java
	* new file

	app/src/main/AndroidManifest.xml
	* add internet permissions

	app/src/main/java/com/example/devbox/stockhawkrewrite/model/IStockDto.java
	app/src/main/java/com/example/devbox/stockhawkrewrite/model/StockDto.java
	* change price-related types from long to float

	app/src/main/java/com/example/devbox/stockhawkrewrite/model/IUtil.java
	* delete interface, not needed

	app/src/main/java/com/example/devbox/stockhawkrewrite/model/Util.java
	* add add method stubs to be implemented

	app/src/test/java/com/example/devbox/stockhawkrewrite/YahooFinanceApiTest.java
	* add BDD tests

+++ app/src/main/java/com/example/devbox/stockhawkrewrite/model/IStockDto.java
@@ -19,11 +19,11 @@ interface IStockDto {
 
     void setmTicker(String mTicker);
 
-    long getmBid();
+    float getmBid();
 
     void setmBid(long mBid);
 
-    long getmAsk();
+    float getmAsk();
 
     void setmAsk(long mAsk);
 
@@ -35,19 +35,19 @@ interface IStockDto {
 
     void setName(String name);
 
-    long getmChangeCurrency();
+    float getmChangeCurrency();
 
     void setmChangeCurrency(long mChangeCurrency);
 
-    long getmChangePercent();
+    float getmChangePercent();
 
     void setmChangePercent(long mChangePercent);
 
-    long getmYearHigh();
+    float getmYearHigh();
 
     void setmYearHigh(long mYearHigh);
 
-    long getmYearLow();
+    float getmYearLow();
 
     void setmYearLow(long mYearLow);
 }
